requirements.txt
flask==3.1.0
flask-CORS==5.0.0
pandas==2.2.3
whoosh==2.7.4
python-dotenv==1.0.1
gunicorn==23.0.0